Secure Chat App/requirements.txt
python
pycryptodome
